+++ CHANGES.rst
@@ -7,6 +7,7 @@ Improvements:
  * Privacy: Make clear that device names are publicly readable (#2662).
  * Widgets: Whitelist [MSC1961](https://github.com/matrix-org/matrix-doc/pull/1961) widget urls.
  * Settings: CALLS section: Always display the CallKit option but grey it out when not available (only on China).
+ * VoIP: Fallback to matrix.org STUN server with a confirmation dialog (#2646).
 
 Changes in 0.9.2 (2019-08-08)
 ===============================================